+++ src/Model/Api/MailProcessorInterface.php
@@ -14,7 +14,7 @@ namespace Fooman\EmailAttachments\Model\Api;
 interface MailProcessorInterface
 {
     public function createMultipartMessage(
-        \Magento\Framework\Mail\MailMessageInterface $message,
+        array $existingParts,
         AttachmentContainerInterface $attachmentContainer
     );
 }